updated baccarat challenge to take advantage of python strings to simplify code and changed outputed plot of results
